# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ajloun Vision Trend Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community feed posts
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    patch:
      operationId: updatePost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostModerationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /posts/{id}/comments:
    get:
      operationId: listComments
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{id}/report:
    post:
      operationId: reportPost
      tags: [posts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ─── OPPORTUNITIES ─────────────────────────────────────
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ─── AUTH ─────────────────────────────
...
    Post:
      type: object
      required: [id, userId, content, likesCount, commentsCount, isApproved, isPinned, isHidden, reportCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        authorNameAr:
          type: string
        authorAvatar:
          type: ["string", "null"]
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isApproved:
          type: boolean
        isPinned:
          type: boolean
        isHidden:
          type: boolean
        reportCount:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        imageUrl:
          type: ["string", "null"]
    PostModerationInput:
      type: object
      properties:
        isApproved:
          type: boolean
        isPinned:
          type: boolean
        isHidden:
          type: boolean
    LikeResult:
      type: object
      required: [liked, likesCount]
      properties:
        liked:
          type: boolean
        likesCount:
          type: integer
    Comment:
      type: object
      required: [id, postId, userId, content, createdAt]
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        authorName:
          type: string
        authorNameAr:
          type: string
        authorAvatar:
          type: ["string", "null"]
        content:
          type: string
        isHidden:
          type: boolean
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string

    # ─── OPPORTUNITIES ────────────────────
...